test: align test files with source files per Go convention
Remove validate_test.go (no matching validate.go) and distribute
its tests to the files they belong to: a_test.go, srv_test.go,
and new common_test.go.
